+++ app/src/main/java/com/yugimaster/jav/UIUtil.java
@@ -1,0 +1,36 @@
+package com.yugimaster.jav;
+
+import android.content.Context;
+import android.util.TypedValue;
+
+/**
+ * This class is for common Android UI.
+ */
+public class UIUtil {
+    /**
+     * Get the height of the screen
+     */
+    public static int getScreenHeight(Context context) {
+        return context.getResources().getDisplayMetrics().heightPixels;
+    }
+    /**
+     * Get the width of the screen
+     */
+    public static int getScreenWidth(Context context) {
+        return context.getResources().getDisplayMetrics().widthPixels;
+    }
+    /**
+     * Dp to Px
+     */
+    public static float dp2px(Context context, float dp) {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp,
+                context.getResources().getDisplayMetrics());
+    }
+    /**
+     * Px to Dp
+     */
+    public static float px2dp(Context context, float px) {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_PX, px,
+                context.getResources().getDisplayMetrics());
+    }
+}